Some things are just not working - play game button, and the disconnect button
