+++ Assets/Scenes/nanobe.unity.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 720e0185087c4d04eb7c89c76bc2028f
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
