+++ Assets/Scenes/watase.unity.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: e71a120f3f6327c44919c6641fda5591
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
