+++ Assets/Scripts/StartMenu/StartMenuManager.cs
@@ -1,20 +1,48 @@
 using System.Collections;
 using System.Collections.Generic;
 using UnityEngine;
+using UnityEngine.UI; // for the toggle
 using Thirdweb; // for interacting with the SDK
 using System.Threading.Tasks; // for some async functionality
+using UnityEngine.SceneManagement;
 
+// this is a combination of Thirdweb tutorial and Watase's Dragon Unity Game example.
 public class StartMenuManager : MonoBehaviour
 {
     // Start is called before the first frame update
     // Create references for the two Game Objects we created
     public GameObject ConnectedState;
     public GameObject DisconnectedState;
+    // public GameObject ConnectWalletState;
+    // public GameObject SelectWallet;
+    public GameObject HasNFTState;
+    public GameObject NoNFTState;
+    public Toggle HonuToggle;
+    public GameObject StartGameButton;
+
     private ThirdwebSDK sdk;
     void Start()
     {
-          // Configure the network you deployed your contracts to:
+      // Configure the network you deployed your contracts to:
       sdk = new ThirdwebSDK("goerli");  
+      DisconnectedState.SetActive(true);
+      ConnectedState.SetActive(false);
+      HasNFTState.SetActive(false);
+      NoNFTState.SetActive(false);
+      // HonuToggle.isOn = false;
+
+    }
+        // Update is called once per frame
+    void Update()
+    {
+      if(HonuToggle.isOn)
+        {
+            StartGameButton.SetActive(true);
+        }
+        else
+        {
+            StartGameButton.SetActive(false);
+        }  
     }
 
     public async void ConnectWallet()
@@ -25,7 +53,7 @@ public class StartMenuManager : MonoBehaviour
               .wallet
               .Connect(new WalletConnection()
               {
-                  provider = WalletProvider.CoinbaseWallet,
+                  provider = WalletProvider.MetaMask,
                   chainId = 5 // Switch the wallet Goerli on connection
               });
               
@@ -42,43 +70,54 @@ public class StartMenuManager : MonoBehaviour
         .GetComponent<TMPro.TextMeshProUGUI>()
         .text = address;
       
-        string balance = await CheckBalance();
+        
+        string balance = await CheckBalance(address);
 
       // Convert the balance to a float
         float balanceFloat = float.Parse(balance);
-
-      // Create text for the balance depending on if the balance is greater than 0
-        string balanceText =
-        balanceFloat > 0
-          ? "Welcome to Happy Honu! Pick a 🐢 to mint:"
-          : "You can't access this without a Key!";
-
-      // Set the ConnectedStates "Balance" GameObjects text to the wallet balance
-        ConnectedState
-          .transform
-          .Find("Owns NFT")
-          .GetComponent<TMPro.TextMeshProUGUI>()
-          .text = balanceText; 
+        print(balanceFloat);
+        if (balanceFloat > 0) {
+            HasNFTState.SetActive(true);
+            print(true);
+          }
+        else {
+            NoNFTState.SetActive(true);
+            print(false);
+          }
     }
-
-    public async Task<string> CheckBalance()
+    public void DisconnectWallet()
     {
-      // Connect to the smart contract
-      // Replace this with your own contract address
-      Contract contract = sdk.GetContract("0xd6343467537998b1EDa668466126cb5b772c2EA1"); // my addy
-      // Contract contract = sdk.GetContract("0x69BC6d095517951Df17f70f38746Bc27CE1C8A62"); // tutorial's addy
-      // Replace this with your NFT's token ID
-      string tokenId = "0";
+      // Disable disconnected state
+      DisconnectedState.SetActive(true);
 
-      // Check the balance of the wallet for this NFT
-      string balance = await contract.ERC1155.Balance(tokenId);
-      return balance;
+      // Enable connected state
+      ConnectedState.SetActive(false);
+      
     }
 
+    public async Task<string> CheckBalance(string address)
+    {
+        Contract contract = sdk.GetContract("0x5e8D21f0bCfD8C57c7dAcE57a2bbd53ad69aa37f");
+
+        string tokenId = "0";
+
+        string balance = await contract.Read<string>("balanceOf", address, tokenId);
+        print(balance);
+        return balance;
+    }
+
+    public async void ClaimHonu()
+    {
+        string address = 
+            await sdk
+            .wallet
+            .Connect(new WalletConnection() {
+                provider = WalletProvider.MetaMask,
+                chainId = 5
+            });
+        Contract contract = sdk.GetContract("0x5e8D21f0bCfD8C57c7dAcE57a2bbd53ad69aa37f");
+        await contract.ERC1155.ClaimTo(address, "0", 1);
+    }
     
-    // Update is called once per frame
-    void Update()
-    {
-        
-    }
+
 }
