--- Assets/Scripts/StartMenuManager.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 43a1c013fff66684190942623381d786
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
