+++ happy-honu.sln
@@ -1,17 +1,20 @@
 ﻿
 Microsoft Visual Studio Solution File, Format Version 11.00
 # Visual Studio 2010
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Assembly-CSharp", "Assembly-CSharp.csproj", "{b81ee5a7-5a95-8dd7-1b57-72ff79fbe3fa}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Assembly-CSharp", "Assembly-CSharp.csproj", "{0a0dc17d-b7d7-8217-783c-7012e547650f}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
 	EndGlobalSection
 	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{b81ee5a7-5a95-8dd7-1b57-72ff79fbe3fa}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{b81ee5a7-5a95-8dd7-1b57-72ff79fbe3fa}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{0a0dc17d-b7d7-8217-783c-7012e547650f}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{0a0dc17d-b7d7-8217-783c-7012e547650f}.Debug|Any CPU.Build.0 = Debug|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
 	EndGlobalSection
 EndGlobal
+
+
+
